Refactor home page to load instantly with parts that need loading in suspense
