+++ components/index.ts
@@ -1,0 +1,3 @@
+export {default as Section} from './Section'
+export {default as Button} from './Button'
+export {default as KofiWidgetOverlay} from './KofiWidgetOverlay'
